Renaming type
